brahma-sync/requirements.txt
acicobra===4.1-2g
acimodel===4.1-2g
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7
